Move `topology::start_validated` into `impl RunningTopology`
